requirements.txt
asgiref==3.6.0
Django==3.2
django-filter==22.1
djangorestframework==3.13.0
djangorestframework-jwt==1.11.0
importlib-metadata==6.0.0
Markdown==3.4.1
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.15.0
